Последняя версия из локального репозитория;
git-svn-id: http://frl.googlecode.com/svn/trunk@22 daf5c975-553e-0410-8cda-3135a9dfd8f3

+++ test/fs_fn/main.cpp
@@ -541,7 +541,7 @@ frl::Bool TestConcat( void )
 	removal( fileName2 );
 	FileOffset length = frl::io::fs::length( fileNameDst );
 	removal( fileNameDst );
-	if( length != text.size() * 2 )
+	if( (size_t)length != text.size() * 2 )
 		return frl::False;
 	return frl::True;
 }
